hardware/requirements.txt
# Requirements for NodeMCU testing script
paho-mqtt==1.6.1
requests==2.31.0
